+++ CHANGELOG.md
@@ -18,6 +18,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `FeedbackBox` and `NewsletterSignup` components to set the `tabindex` on the internal elements to `"-1"`.
 - Updates `DatePicker`'s `TextInput` to always have an `aria-label` attribute that tells screen reader users how to access the calendar.
 
+### Fixes
+
+- Fixes a bug in the `Hero` component where custom background colors were not rendering properly in dark mode for the `"campaign"` and `"tertiary"` variants.
+
 ### Deprecates
 
 - Deprecates the `"secondary"` and `"fiftyFifty"` variants of the `Hero` component.